requirements.txt
requests
peewee
PyMySQL[rsa]
pymongo
toml
attrs
loguru
python-dotenv
numpy<2
opencv-python
portkey-ai
arrow
--extra-index-url https://gitlab.com/api/v4/projects/56789260/packages/pypi/simple
geetest-captcha-v3-icon-analysis
DrissionPage
websocket-client
uiautomator2